==> 1 - Data Collection - Housing Address Coordinates.ipynb <==
# Singapore Public Housing (HDB) Resale Price Prediction Model (Part 1)
### Data Collection - Housing Address Coordinates

## 1. Introduction

In the late 19th century, American steel mogul Andrew Carnegie famously made a bold prediction that investment in real estate would be the wise way to multiply one's wealth. This statement has turned out to be relevent and agreeable to most of us 100 years later in the 21th century.

>“Ninety percent of all millionaires become so through owning real estate. More money has been made in real estate than in all industrial investments combined. The wise young man or wage earner of today invests his money in real estate.” - Andrew Cargie 

According to Forbes rich list in 2018, out of the 22 billionaires in Singapore, 16 of them are real-estate tycoons. The total fortune came up to a staggering $43.7 billion, which grew up to nearly 10 percent from the previous year (Channel News Asia, 2018). We cannot deny that all the successful real-estate investors have had either natural talent or decade-trained experience in them to tell a good or bad property listing or development project apart. Nevertheless, it would be fascinating if we could unravel and quantify the investors' - or ordinary home-buyer - judgement with the help of Data Science. So we set out a journey to create a machine learning model that could benchmark a property pricing based on its attributes.

In this project, we will develop a predictive model on HDB unit resale price using several machine learning techniques of different classes. We believe that a model like this would be very valuable for any real state agent or home-buyer to set a benchmark pricing to judge whether a property listing is over or under-valued.

One thing worth noting is that HDB-developed flats are in a highly controlled environment by the Government to avoid market speculation on housing price. In fact, investment in these flats is technically prohibited as one could only own a maximum of one HDB flat. However, this situation does not render our study and model useless, simply because of the fact that this project still serve as a window to look into home buyer's psychology and how premium is justifiable by numerical features or attributes. So the same methodology could still be applied to Condominium sales in Singapore or real estates in City States that resembles Singapore where some of the key features could still be applicable and worth researching.

## 2. Problem Statement

Singapore has always had one of the most expensive housing market in the world, so it becomes crucial for the locals to make sure every dollar they spent are worthwhile. However, most of the housing price are still benchmarked manually by experienced appraiser today. Hence, it would really be helpful for home buyer if there is a predictive model available to find out any undervalued property listing and maximize their savings.

## 3. About this Notebook

The raw data for the project is hailing from the official Singapore Government portal, where the archive can be found in [this link](https://data.gov.sg/dataset/resale-flat-prices). It is transaction record for all resale public housing flat under the development and management of Housing Development Board (HDB) in Singapore. These flats are commonly known as "HDB" by the locals in the Singapore. 

The transaction data available range from 1990 to 2020, housing price has rose significantly through the last few decades and the property price inflation would not be well captured by the model without thorough research through public policies and global economic climate. Hence, the project will only encapsulated property sold from January 2017 to March 2020.

The raw data has come with some basic information on the property attributes, such as the situated town, storey range, floor area, remaining lease and the resale price (the target variable). All HDB flats are bound by a standard 99-year lease before government has the right to seize the property for future development, so it is an important indicator of housing price in Singapore.

In this notebook, we will clean up and transform the necessary raw data to its numerical format for the ease of modelling and visualization. Then, with the use of OneMap public API, we will submit each address and transform each data entries into their respective coordinates so that we can input more geospatial data in the later stage. 

## 4. Initialization

# Import Vanilla Libraries
import requests, json, time, random, math
import pandas as pd
import numpy as np

# Read property dataset as dataframe
hdb = pd.read_csv('./Dataset/Raw/hdb_2017_2020.csv')

hdb.head()

## 5. Feature Engineering

A quick glance through the first 5 rows of data reveals that some of the data (such as month and remaining lease) are not in a usable format. We will need to extract and transform these data into numerical form.

### 5.1 Date Time Data

# Extract Month and Year from 'month' column
hdb['sold_year'] = hdb['month'].str[:4].astype(int)
hdb['sold_month'] = hdb['month'].str[-2:].astype(int)

# Function to convert 'remaining lease' to numerical column
def extract_lease_years(row):
    if len(row)==4:
        return round(int(row[0]) + int(row[2])/12, 2)
    else:
        return int(row[0])

# Apply function to 'remaining lease'
hdb['remaining_lease'] = hdb['remaining_lease'].str.split().apply(extract_lease_years)

hdb.sample(5)

After transforming the date into month and year and remaining lease into float number, we would need to concatenate the block number to its respective street name so that we can pass the full address into OneMap API for more accurate response of the coordinates.

# Concatenate block number and street name to form full address for API Calls
hdb['address'] = hdb['block'] + ' ' + hdb['street_name']

# Drop all the pre-engineered column
hdb.drop(['month', 'block', 'street_name'], axis=1, inplace=True)

### 5.2 Storey Range

For the storey range, we will simply extract the median of each range; since the range only covers 3 levels, the data loss would be almost negligible.

# Check on storey range
hdb.storey_range.unique()

# Convert 'storey_range' into numerical by taking the midpoint of each range
hdb.storey_range = (hdb.storey_range.str[:2].astype(int) + hdb.storey_range.str[-2:].astype(int))
hdb.storey_range = (hdb.storey_range/2).astype(int)

# Sanity check
hdb.head()

### 5.3 Coordinates

#### 5.3.1 OneMap API Calling

In order to speed up the API calling, instead of passing each entries, which requires 70,104 requests, we will only pass in unique address we have in the raw dataset. We have only 8,712 unique address, so the time required for API calling has been saved by 88%.

unique_add = pd.Series(hdb['address'].unique())

unique_add = pd.DataFrame(unique_add, columns=["address"])

print("Time Saving on API Calling:", round(1 - unique_add.shape[0] / hdb.shape[0], 2) * 100, '%')

# Extract coordinates from OneMap API by sending full address of each block
latitude = []
longitude = []

if 'latitude' not in unique_add.columns:
    unique_add['latitude'] = np.nan
    unique_add['longitude'] = np.nan

for i, address in enumerate(unique_add['address']):
    if math.isnan(unique_add.loc[i, 'latitude']):
        try:
            print('\rWaiting... ({})... {} addresses remaining... '.format(address, len(unique_add)-i-1), end='.')
            query = "https://developers.onemap.sg/commonapi/search?searchVal=" + address + "&returnGeom=Y&getAddrDetails=N"
            response = requests.get(query)
            coor_json = json.loads(response.content)
            unique_add.loc[i, 'latitude'] = coor_json['results'][0]['LATITUDE']
            unique_add.loc[i, 'longitude'] = coor_json['results'][0]['LONGITUDE']

        except:
            unique_add.loc[i, 'latitude'] = np.nan
            unique_add.loc[i, 'longitude'] = np.nan

        # Export dataframe every loop to csv for inspection
        unique_add.to_csv('./Dataset/Transitional/unique_address.csv', index=False)

        # Sleeping time to avoid overloading server
        time.sleep(random.randint(1,2)/4)

if unique_add['latitude'].isnull().sum() == 0:
    print("--- Data is complete ---")

#### 5.3.2 Addressing Missing Data

It seems that we have a systematic pattern of missing data from the API calling. All entries have full address containing the word "St. George's". A quick trial on OneMap has showed that the system only register the word 'Saint' rather than the abbreviation of 'St.'. So we will replace all the abbreviation and pass it again into the loop. But this time, the function will be updated with a NaN checking feature, so only missing value will be updated, saving us previous time and resources.

unique_add[unique_add['latitude'].isnull()]['address'].unique()

missing_index = unique_add[unique_add['latitude'].isnull()].index

unique_add.loc[missing_index, 'address'] = unique_add.loc[missing_index, 'address'].str.replace('ST.', 'SAINT')

unique_add['latitude'] = unique_add['latitude'].astype(float)
unique_add['longitude'] = unique_add['longitude'].astype(float)

#### 5.3.3 Second Round API Calling

# Repassing dataframe for API Calling but only missing data is processed
for i, lat in enumerate(unique_add.latitude):
    
    # Only if latitude is missing, fire up API calling query
    if math.isnan(lat):
        address = unique_add.loc[i, 'address']
        try:
            print('\rWaiting... ({})... {} neighbourhood remaining... '.format(address, len(unique_add)-i-1), end='.')
            query = "https://developers.onemap.sg/commonapi/search?searchVal=" + address + "&returnGeom=Y&getAddrDetails=N"
            response = requests.get(query)
            coor_json = json.loads(response.content)
            unique_add.loc[i, 'latitude'] = coor_json['results'][0]['LATITUDE']
            unique_add.loc[i, 'longitude'] = coor_json['results'][0]['LONGITUDE']
            time.sleep(random.randint(1,2)/4)

        except:
            unique_add.loc[i, 'latitude'] = np.nan
            unique_add.loc[i, 'longitude'] = np.nan
            time.sleep(random.randint(1,2)/4)
            
unique_add.to_csv('./Dataset/Transitional/unique_address.csv', index=False)

if unique_add['latitude'].isnull().sum() == 0:
    print("--- Data is complete ---")

# Sanity check on missing data
unique_add = pd.read_csv("./Dataset/Transitional/unique_address.csv")

unique_add.isnull().sum()

## 6. Final Data Export

hdb = pd.merge(hdb, unique_add, how='left', on='address')

# Export final dataframe to CSV
hdb.to_csv('./Dataset/Transitional/complete_data.csv', index=False)

# Sanity check on the final dataframe
hdb = pd.read_csv('./Dataset/Transitional/complete_data.csv')
hdb.isnull().sum()

missing_data = hdb[hdb['latitude'].isnull()].index
hdb.loc[missing_data, 'address'] = hdb.loc[missing_data, 'address'].str.replace('ST.', 'SAINT')
hdb = pd.merge(hdb, unique_add, how='left', on='address')

After the data from unique address is merged to the final data from the first round, we now have 2 sets of coordinates. So we will be dropping the data from the first round as they are encompassed in the second round data as well. The 2 columns will also be renamed for the ease of future references.

hdb.isnull().sum()

hdb.drop(['latitude_x', 'longitude_x'], axis=1, inplace=True)
hdb.rename(columns={'latitude_y': 'latitude', 'longitude_y': 'longitude'}, inplace=True)

# Export final dataframe to CSV
hdb.to_csv('./Dataset/Transitional/complete_data.csv', index=False)

Now that the final data has been checked and exported, we will move on the second part on data scraping. In the 2nd notebook, data on amenities and infrastructure will be scraped from online resource and merged with our HDB dataset so that the effect of each aspects on housing price can be studied.
