==> 2 - Data Collection - Amenities and Infrastructure Data.ipynb <==
# Singapore Public Housing (HDB) Resale Price Prediction Model (Part 2)
### Data Collection - Amenities and Infrastructure Data

## 1. About this Notebook

It is common sense that property prices are heavily dictated by its geographical location apart from its attributes. This phenomenon is especially true in Singapore where most of its people travel by their extensive network of public transport system. Proximity to amenities such as schools, malls and to infrastructures such as bus terminal and MRT (Mass Rapid Transit) train station becomes crucial factors when home buyers are looking out for listings.

So in this notebook, we will be performing web scraping and API calling to collect data from online resources such as Wikipedia. After gathering these data on amenities and infrastructure in Singapore, we will merge these data to the core HDB data to further study the effect of each of these factors. Same as the notebook previously, OneMap API will be used to retrieve the coordinates of each place of interest so that we can map these data to the core data by its proximity.

There are 5 types of venues that we will be collecting for this part of the project, namely:
1. MRT Station - its location and whether it is MRT interchange or Bus interchange or neither
2. Bus Stops - its location
3. Primary School - its vacancy count, location and whether it has affiliation with any Secondary School
4. Secondary School - its entry cut-off points and whether it has affiliation with any Junior College
5. Shopping Mall - its location.

## 2. Initialization

# Import Vanilla Libraries
import requests, json, time, random, re
import pandas as pd
import numpy as np

# Import Library for Web Scraping
from bs4 import BeautifulSoup

To our codes DRY, we will be using a main coordinate retrieval function that submit calls to OneMap API and collect the latitude and longitude of each entry. Variable sleeping time has been designed in the function to prevent server overloading. Any entry that returns no coordinates will be assigned as NaN and will be addressed as per each situation.

# MAIN COORDINATES RETRIEVAL FUNCTION

def coordinate_retrieval(df_name, target_col, mrt=False):
    latitude = []
    longitude = []

    for i, entry in enumerate(df_name[target_col]):
        try:
            print('\rWaiting... ({})... {} entries remaining... '.format(entry, len(df_name)-i-1), end='.')
            if mrt:
                query = "https://developers.onemap.sg/commonapi/search?searchVal=" + entry + ' mrt' + "&returnGeom=Y&getAddrDetails=N"
            else:
                query = "https://developers.onemap.sg/commonapi/search?searchVal=" + entry + "&returnGeom=Y&getAddrDetails=N"
            response = requests.get(query)
            coor_json = json.loads(response.content)
            latitude.append(coor_json['results'][0]['LATITUDE'])
            longitude.append(coor_json['results'][0]['LONGITUDE'])

        except:
            latitude.append(np.nan)
            longitude.append(np.nan)

        time.sleep(random.randint(1,2)/4)
    
    return latitude, longitude

## 2. Data Scraping

### 2.1 Public Transport Data

#### 2.1.1 Mass Rapid Transit (MRT)

The MRT data will be acquired through web scraping on Wikipedia. The page has an up-to-date table summarizing all the existing and future train station in Singapore, the website can be found [here](https://en.wikipedia.org/wiki/List_of_Singapore_MRT_stations). Some data cleaning are required to extract the information that we need from this page, for example removing future station that has not been built yet and avoid duplicates of train station that are interchange for 2 or more lines.

# Scraping table from wikipedia
url = "https://en.wikipedia.org/wiki/List_of_Singapore_MRT_stations"
res = requests.get(url)
soup = BeautifulSoup(res.content,'lxml')
table = soup.find_all('table')[1]

# Convert table into dataframe
dfs = pd.read_html(str(table))
mrt = dfs[0]
mrt

# Drop off first level of header
mrt.columns = mrt.columns.droplevel(0)
mrt.head()

# Drop off secondary header on MRT Line Name 
# by comparing Chinese and Tamil Station Name
mrt = mrt[mrt['Chinese'] != mrt['Tamil']]
mrt.head()

# Drop off Unnecessary Columns
mrt = mrt.drop(['Chinese', 'Tamil', 'Opening', 'Name(s) during planning stages', 'Abbreviation', 'Location(s)'], 
             axis=1)
mrt.head()

# Renaming columns to be more readable
mrt.columns = ['Code', 'Future Code', 'Name', 'Interchange']
mrt.head()

# Drop off Station under construction
mrt = mrt.dropna(subset=['Code'])

# Drop off duplicates
# Interchange Stations that shows up more than 1 time

mrt = mrt.drop_duplicates(subset='Name')

# Fill up empty cells in Interchange column with empty string
mrt['Interchange'] = mrt['Interchange'].fillna(' ')

There are some MRT station that has the role of being a Bus Interchange as well. So as long the word 'Bus Interchange' is mentioned in the Interchange column, the station will be classified as 1 for positive.

As for MRT interchange, if there are more than 1 station code in the Code column, it would signify that the MRT station is an interchange to the other train line/service. Thus the station will be classified as 1 for positive.

# Function to return boolean code on whether 
# the MRT station is also a bus interchange

def bus_interchange(element):
    if 'Bus Interchange' in element:
        return 1
    else:
        return 0

mrt['Bus_Interchange'] = mrt['Interchange'].apply(bus_interchange)

# Function to return boolean code on whether 
# the MRT station is a MRT interchange

def mrt_interchange(element):
    
    # If there are more than 1 code for a single station
    if len(element.split())>1:
        return 1
    
    else:
        return 0
    
mrt['MRT_Interchange'] = mrt['Code'].apply(mrt_interchange)

# Drop off pre-engineered column
mrt = mrt.drop(['Code', 'Future Code', 'Interchange'], axis=1)

# Fixing "Botanic Gardens" name
mrt['Name'].replace({
    'Botanic Gardens • Kebun Bunga' : 'Botanic Gardens',
}, inplace=True)

# Reset Index for dataframe
mrt.reset_index(drop=True, inplace=True)

# Extract coordinates for each MRT station for distance calc
mrt['Latitude'], mrt['Longitude'] = coordinate_retrieval(mrt, "Name", mrt=True)

# The only null column is a LRT station
mrt[mrt['Latitude'].isnull()]

mrt.dropna(inplace=True)

mrt.sample(5)

# Export data to CSV
mrt.to_csv('./Dataset/Engineered/MRT.csv', index=False)

#### 2.1.2 Bus Stops

Work on scraping bus stops data in Singapore have been done by Mengyong Lee in January 2019 with the use of OneMap API, you may find his article **"Data-Driven Approach to Understanding HDB Resale Prices in Singapore"** on [towards data science](https://towardsdatascience.com/data-driven-approach-to-understanding-hdb-resale-prices-in-singapore-31c3beecfd97) if you are interested. The data was last updated in February 2020, so it is up to date for our study of HDB transaction up to March 2020. His code and dataset can be found in [this github repositories](https://github.com/mylee16/onemap-api).

Minimal work needed to be done on the dataset as they have already contain the coordinates for each bus stop entries.

# Read data from csv file
bus_stop = pd.read_csv('./Dataset/Raw/bus_stops.csv')

# Check on the data
bus_stop.head()

# Dropping unnecessary columns
bus_stop.drop(['Unnamed: 0', 'RoadName', 'BusStopCode'], axis=1, inplace=True)

# Renaming columns to be more readable
bus_stop.columns = ['name', 'latitude', 'longitude']

# Sanity check on dataframe
bus_stop.sample(5)

# Export data to CSV
bus_stop.to_csv('./Dataset/Engineered/Bus_Stop.csv', index=False)

### 2.2 Education

For both the primary school and secondary school, the data will be scraped from Salary.sg, a blogspot with content focusing on local news, wealth management and investments in Singapore. To avoid unhealthy competition between schools, there is no official ranking for any public education institution in Singapore. So to quantify the elite school, we will be using other metrics:

For primary school, we will be looking at the vacancy count, which is a good proxy for how popular each school is as seen by the public. The lower the vacancy available usually indicates how competitive is for students to enroll into the said school. The link to the article can be found [here](https://www.salary.sg/2019/best-primary-schools-2019/).

While for secondary school, the metric to measure or define an elite school would be accessing its cut-off point, whereby any student with score higher than the cut-off would not be granted a place in the school. Thus a lower cut-off point would signify that the school has higher requirement for the intellectual prowess of their student. You may find the link to the article [here](https://www.salary.sg/2018/secondary-school-ranking-based-on-cut-off-for-2018-intake/).

#### 2.2.1 Primary School

# Scraping data from website with primary school detail
url = "https://www.salary.sg/2019/best-primary-schools-2019/"
res = requests.get(url)
soup = BeautifulSoup(res.content,'lxml')
lis = soup.find('ol')

# List with all schools
all_ps = [li.text for li in lis.find_all('li')]

# List with only school with affiliation
aff_ps = [strong.text for strong in lis.find_all('strong')]

# Transform scraped data into dataframe
ps_df = pd.DataFrame(all_ps, columns=['name'])
ps_df.head()

# Extract vacancy data
ps_df['vacancy'] = [row[-1] for row in ps_df['name'].str.split()]

# Random check on vacancy
ps_df.sample(5)

# Function to create new column 'affiliation'
def aff_school(row):
    if row in aff_ps:
        return 1
    else:
        return 0
    
ps_df['affiliation'] = ps_df['name'].apply(aff_school)

# Remove vacancy value and symbols from "name"
ps_df['name'] = ps_df['name'].str.extract(r'(.+) –.+')

# Function to remove all secondary information in name
def extract_name(row):
    if '(' in row:
        return re.findall(r'(.+) \(', row)[0]
    else:
        return row

# 1st-loop to remove one parentheses (boy/girl school)
ps_df['name'] = ps_df['name'].apply(extract_name)

# 2nd-loop to remove second parentheses (string of "Primary")
ps_df['name'] = ps_df['name'].apply(extract_name)

# Rectify name to OneMap standard address naming
ps_df['name'] = ps_df['name'].str.replace('’', "'").str.replace(r'St\.', 'Saint')

# Sanity check on affiliation
ps_df.sample(5)

# Extract coordinates for distance calc
ps_df['latitude'], ps_df['longitude'] = coordinate_retrieval(ps_df, "name")

# Missing rows = None
ps_df[ps_df['latitude'].isnull()]

ps_df.head(10)

# Export data to CSV
ps_df.to_csv('./Dataset/Engineered/Primary_School.csv', index=False)

#### 2.2.2 Secondary School

# Scraping data from website with secondary school detail
url = "https://www.salary.sg/2018/secondary-school-ranking-based-on-cut-off-for-2018-intake/"
res = requests.get(url)
soup = BeautifulSoup(res.content,'lxml')

lis = soup.find('ol')

# List with all schools
all_ss = [li.text for li in lis.find_all('li')]

# List with only school with affiliation
aff_ss = [strong.text for strong in lis.find_all('strong')]

# Transform scraped data into dataframe
ss_df = pd.DataFrame(all_ss, columns=['name'])
ss_df['cutoff_point'] = [row[-1] for row in ss_df['name'].str.split()]
ss_df.sample(5)

# Function to create new column 'affiliation'
def aff_school(row):
    if row in aff_ss:
        return 1
    else:
        return 0
    
ss_df['affiliation'] = ss_df['name'].apply(aff_school)

# Remove vacancy value and symbols from "name"
ss_df['name'] = ss_df['name'].str.extract(r'(.+) –.+')

# Function to remove all secondary information in name
def extract_name(row):
    if '(' in row:
        return re.findall(r'(.+) \(', row)[0]
    else:
        return row

# 1st-loop to remove one parentheses (co-ed)
ss_df['name'] = ss_df['name'].apply(extract_name)

# 2nd-loop to remove second parentheses (string of "Secondary")
ss_df['name'] = ss_df['name'].apply(extract_name)

# Rectify name to OneMap standard address naming
ss_df['name'] = ss_df['name'].str.replace('’', "'").str.replace(r'St\.', 'Saint')

# Rectify name to remove the programme name
ss_df['name'] = ss_df['name'].str.replace('IP', '').str.replace('SAP', '')
ss_df['name'] = ss_df['name'].str.replace('O-levels', '').str.replace('IB', '')
ss_df['name'] = ss_df['name'].str.replace('Govt.', 'Government')
ss_df['name'] = ss_df['name'].str.replace('–', '').str.strip()

# Extract coordinates for distance calc
ss_df['latitude'], ss_df['longitude'] = coordinate_retrieval(ss_df, "name")

# Missing rows
ss_df[ss_df['latitude'].isnull()]

According to research online, these 2 schools have been closed down by the Singapore government for merging with the other schools to improve the productivity and effectiveness. So we will have to drop these rows from the dataset.

# Remove closed-down school
ss_df = ss_df.dropna(subset=['latitude'])

# Export final dataframe to csv
ss_df.reset_index(drop=True).to_csv('./Dataset/Engineered/Secondary_School.csv', index=False)

## 2.3 Shopping Mall

Similar to the MRT station data collection, shopping malls data can be found in Wikipedia where they are summarized and categorized into the area where it is situated at. The webpage can be found in this [link](https://en.wikipedia.org/wiki/List_of_shopping_malls_in_Singapore).

# Making request to wikipedia page
url = "https://en.wikipedia.org/wiki/List_of_shopping_malls_in_Singapore"
res = requests.get(url)
soup = BeautifulSoup(res.content,'lxml')

# Scraping through all 7 tables in wikipedia page
mall_df = []
for idx in range(1, 8):
    mall_df += [lis.text for lis in soup.find_all('ul')[idx].find_all('li')]

# Rectify mall name to remove all symbols
mall_df = [re.findall(r'(.+)\[', mall)[0].strip() if '[' in mall else mall for mall in mall_df]
mall_df = [re.findall(r'(.+)\(', mall)[0].strip() if '(' in mall else mall for mall in mall_df]
mall_df = [re.findall(r'(.+),', mall)[0].strip() if ',' in mall else mall for mall in mall_df]
mall_df = [re.findall(r'(.+)@', mall)[0].strip() if '@' in mall else mall for mall in mall_df]

# Convert list into a dataframe
mall_df = pd.DataFrame(mall_df, columns=['name'])

# Extract coordinates for distance calc
mall_df['latitude'], mall_df['longitude'] = coordinate_retrieval(mall_df, "name")

# Malls with discrepancy with actual name
mall_df[mall_df['latitude'].isnull()]

There are several rows that return no result because the name does not match exactly to the official name of the shopping mall. However, they attributed to only less than 10% of the data collected, we may drop the rows from the dataset. 

# Drop off missing rows
mall_df.dropna(inplace=True)

# Export data to CSV
mall_df.reset_index(drop=True).to_csv('./Dataset/Engineered/Mall.csv', index=False)

With the infrastructure data ready in CSV format, we will continue to gather geospatial data in the next part of the project. In the next notebook (Part 3), we will be looking into travel time to 5 different employment center or hotspot in Singapore to study the effect of each location.
