==> 6 - Feature Engineering - Hexagon Map.ipynb <==
# Singapore Public Housing (HDB) Resale Price Prediction Model (Part 6)
### Feature Engineering - Hexagon Map

## 1. About this Notebook

There is no way to visualize the distribution of sale price geographically better than plotting a heatmap over a Singapore streetmap. The heatmap could show the hotspot in each township, the effect of amenities and infrastructure and probably factors that we have not thought of yet. After a few trials on different technique available in Python, it came to conclusion that plotting the distribution as a choropleth map using Folium is the easiest to customize and without a need to API key/token.

This notebook will not covers the plotting part, instead we will now focus on the feature engineering and wrangling of geospatial data using Geojson file. The original raw file was downloaded from [Singapore online archive](https://data.gov.sg/dataset/master-plan-2019-planning-area-boundary-no-sea), where the planning area (similar to township) are plotted into Geojson format. However, there is a minor discrepancy in the planning area (zoned by Urban Development Board, URA) and the actual HDB township. So work was done away from the notebook to add on one more sub-area of Whampoa, which will be merged with Kallang to form the HDB township of Kallang/Whampoa. The original planning area had Whampoa included in Novena instead of Kallang.

The general approach is to divide the Singapore map into equal size of hexagon. Then we would label each HDB address into their situated hexagon and obtain a mean value for each hexagon. As such, a heatmap can be plotted with an choropleth overlay with the feature that we want to study in accordance to granularity we prefer. For this project, each hexagon is set to be 250 meters in diameter; however, do feel free to change the diameter in function create_hex_map if you prefer other granularity.

Part of this notebook was performed on Google Colab due to known dependencies issues with Geopandas. The code involving geopandas Geojson driver have been commented out for the ease of running the notebook without error. You may uncomment the code if you would like to test the code.

## 2. Initialization

# # Dependencies for Hexagon-Map Creation
# # No need to import
# !apt install libspatialindex-dev
# !pip install rtree, geopandas

# Vanilla packages
import numpy as np
import pandas as pd
import math

# Geospatial Libraries
import geopandas as gpd
from shapely.geometry import Point
from shapely import wkt
from matplotlib.patches import RegularPolygon

pd.set_option('max_columns', 99)

hdb = pd.read_csv('./Dataset/final_data.csv')

## 3. Hexagon Map Data Preparation

Similar to the naive-haversine function that was created previously in Part 5, the function below is the actual haversine formula that is more accurate than the previous one. It requires more computing power so it was not used previously. However, in this application, we need all the hexagons to line up edge to edge so accuracy is paramount.

# Function to calculate haversine distance between two coordinates
def haversine(coord1, coord2):
    lon1, lat1 = coord1
    lon2, lat2 = coord2
    
    # Earth radius
    R = 6_371_000
    phi_1 = np.radians(lat1)
    phi_2 = np.radians(lat2)
    delta_phi = np.radians(lat2 - lat1)
    delta_lambda = np.radians(lon2 - lon1)
    a = np.sin(delta_phi / 2.0) ** 2 + np.cos(phi_1) * np.cos(phi_2) * np.sin(delta_lambda / 2.0) ** 2
    c = 2 * np.arctan2(np.sqrt(a),np.sqrt(1 - a))
    meters = R * c
    km = meters / 1000.0
    meters = round(meters)
    km = round(km, 3)
    return meters

The code below is to remove planning area that was planned for any HDB development and to combine several planning area into a HDB township. By removing the area that has no HDB at all, we could save up valueable processing power and time computing for hexagon. As mentioned before, Kallang and Whampoa have to be merged to form the HDB township of Kallang/Whampoa. Apart from that, the central area of Singapore divided up into multiple planning area for development efficiency; for our application, we will merged these area into a HDB township of Central Area.

# ### ---      KNOWN DEPENDENCIES ISSUE      --- ###
# ### ---  CODES TO EXPORT FOR GOOGLE COLAB  --- ###
# ### ---     UNCOMMENT TO TEST THE CODE     --- ###

# # Loading data using Geopandas
# df = gpd.read_file('./Raw/planning_area_with_whampoa.json')

# # Drop off planning area that are not HDB estate to save up processing power and memory
# df = df.drop([0, 1, 2, 3, 4, 5, 7, 8, 10, 11, 13, 14, 15, 16, 18, 19, 25, 26, 29, 30, 34, 39, 44, 51, 53]).reset_index(drop=True)

# # Combining smaller planning areas in accordance of HDB estate neighbourhood in raw DataFrame
# df.loc[[0, 25, 26, 29, 30], 'name'] = 'CENTRAL AREA'
# df.loc[[1, 31], 'name'] = 'KALLANG/WHAMPOA'

# # Dissolve area with the same name
# TO_boundary = df.dissolve(by='name')

# # Export to geojson file for further processing
# TO_boundary.to_file('./Dataset/Spatial/planning_area_optimized.geojson', driver='GeoJSON')

# Function to create hexagon polygon on geojson file
def create_hex_map(hex_diam=250):
    # Import the optimized planning area polygons
    df = gpd.read_file('./Dataset/Spatial/planning_area_optimized.geojson')

    # Furthest coordinates on Singapore
    xmin, ymin, xmax ,ymax = df.total_bounds

    # East-West & North-South Length of Singapore
    EW = haversine((xmin, ymin), (xmax, ymin))
    NS = haversine((xmin, ymin), (xmin, ymax)) + 10000

    # Diameter of hexagon in meters
    d = hex_diam

    # Calculate width of hexagon in meters and grid structure of hexagon
    w = d * np.sin(np.pi / 3)
    n_cols = int(EW / w) + 1
    n_rows = int(NS / d) + 1

    # Calculate width of hexagon in coordinates
    w = (xmax - xmin) / n_cols

    # Calculate diameter of hexagon in coordinates
    d = w / np.sin(np.pi / 3)

    array_of_hexes = []

    # Creating multipolygon for each hexagon
    for rows in range(0, n_rows):
        hcoord = np.arange(xmin, xmax, w) + (rows % 2) * w / 2
        vcoord = [ymax - rows * d * 0.75] * n_cols
        for x, y in zip(hcoord, vcoord):
            hexes = RegularPolygon((x, y), numVertices=6, radius=d/2, alpha=0.2, edgecolor='k')
            verts = hexes.get_path().vertices
            trans = hexes.get_patch_transform()
            points = trans.transform(verts)
            array_of_hexes.append(Polygon(points))

    # Transforming multipolygon into geo-dataframe and export to Geojson & CSV
    hex_grid = gpd.GeoDataFrame({'geometry': array_of_hexes}, crs={'init': 'epsg:4326'})
    TO_hex = gpd.overlay(hex_grid, df)
    TO_hex = gpd.GeoDataFrame(TO_hex,geometry='geometry')
    TO_hex.to_file('./Dataset/Spatial/planning_area_hexed.geojson', driver='GeoJSON')
    TO_hex[['geometry']].to_csv('./Dataset/Spatial/HexMap.csv')

# ### ---      KNOWN DEPENDENCIES ISSUE      --- ###
# ### ---  CODES TO EXPORT FOR GOOGLE COLAB  --- ###
# ### ---     UNCOMMENT TO TEST THE CODE     --- ###

# create_hex_map()

# Load in hex-ed singapore map into Geopandas DataFrame
hex_map = pd.read_csv('./Dataset/Spatial/HexMap.csv')
hex_map['geometry'] = hex_map['geometry'].apply(wkt.loads)
hex_map = gpd.GeoDataFrame(hex_map, geometry='geometry')

## 4. Feature Engineering

Next up, we will perform some feature engineering to study the true effect of amenities and other features. We will create be creating 3 different type of metrics:

1. **price_per_sqft** - The simplest form of unit price in the real estate market
2. **unit_price** - It is the price_per_sqft divided by the remaining year lease, eliminating the effect of property lease
3. **elem_price** - Price metric that is more elementary than the unit_price where it has factored in the effect of storey range as well

We would also remove Terrace HDB houses from the dataset as they proved to be outliers as contrary to the general HDB flats. They are extremely rare in Singapore, in fact only 2 neighbourhoods are captured in the dataset. The inclusion would have negligible effect on the model accuracy as the sample size is too low.

# Remove terrace-house (outliers) from dataset for better mapping visualization
hdb = hdb[hdb['flat_model']!='Terrace'].reset_index(drop=True)

# Price per square feet
hdb['price_per_sqft'] = hdb['resale_price'] / hdb['floor_area_sqm'] * 10.7639

# Unit Price --- Price per area per year lease
hdb['unit_price'] = hdb['price_per_sqft'] / hdb['remaining_lease']

# Elementary Price --- Price per area per year lease per storey level
hdb['elem_price'] = hdb['unit_price'] / hdb['storey_range']

# # Original method of grouping property entries by address
# hex_df = hdb.groupby('address').mean().reset_index()

# Renewed method to bypass calculation running
hex_df = pd.read_csv('./Dataset/final_data_with_hex.csv')

# Label each address to their respective hexagon
if hex_df['hex_id'].isnull().sum() > 0:
    for idx, (latitude, longitude) in enumerate(zip(hex_df['latitude'], hex_df['longitude'])):
        print('\rWaiting... {} addresses remaining... '.format(len(hex_df)-idx-1), end='.')
        if math.isnan(hex_df.loc[idx, 'hex_id']):
            for boundary, hex_id in zip(hex_map.geometry, hex_map.hex_id):
                coor = Point(longitude, latitude)
                if coor.within(boundary):
                    hex_df.loc[idx, 'hex_id'] = hex_id
                    break
            hex_df.to_csv('./Dataset/final_data_with_hex.csv', index=False)

else:
    print('--- DATA IS COMPLETE ---')

We will further optimize the data by taking out all hexagon that has no data entry at all. It would cut down the processing time during mapping significantly, so we could zoom in and out the map with ease.

# Further Optimize by taking only hexagon with address/data
hex_map.columns = ['hex_id', 'geometry']
hex_map = hex_map[hex_map['hex_id'].isin(hex_df['hex_id'].unique())]

# ### ---      KNOWN DEPENDENCIES ISSUE      --- ###
# ### ---  CODES TO EXPORT FOR GOOGLE COLAB  --- ###
# ### ---     UNCOMMENT TO TEST THE CODE     --- ###

# # Export for Colab
# hex_map.to_csv('./Dataset/Spatial/HexMap_Optimized.csv', index=False)

# # Read from Colab
# hex_map = pd.read_csv('./Dataset/Spatial/HexMap_Optimized.csv')
# hex_map['geometry'] = df['geometry'].apply(wkt.loads)
# hex_map = gpd.GeoDataFrame(df, geometry='geometry')

# # Export the optimized hexmap to geojson format for mapping purpose
# hex_map.columns = ['hex_id', 'geometry']
# hex_map.to_file('./Dataset/Spatial/hexmap.geojson', driver='GeoJSON')

# Remove any missing points not hex-ed (outliers far from actual HDB estate)
# Only 3 addresses affect - 1 x Changi Village + 2 x Farrer Road
hex_df = hex_df.dropna().reset_index(drop=True)

# Only merged if column not exist - to ease re-running of code
if 'town' not in hex_df.columns:
    hex_df = pd.merge(hex_df, hdb[['address', 'town']], how='left', on='address')

# Merging cause dataframe to unpack, drop all duplicates with the same address
hex_df = hex_df.drop_duplicates().reset_index(drop=True)

Due to the general trend where property is usually more expensive when they are closer to the city center, we would not be able to study the heatmap at suburban of Singapore as they will all be at the lower rung for all price metrics. Increasing the granularities would cause more harm than good on area-centric study. So one of the method that we have adopted is to calculate a township-wide price index.

It is dubbed as 'Appeal Score' in this project, where it is a measure of how property appeal to home buyer at its township. Mathematically, it is unit price (resale price but stripped away the effect of lease and floor area) that is normalized to a scale of 0 to 1 based on the local minimum and maximum.

# Aggregate HDB-town-wide local Minimum and Maximum
town_ave = hex_df.groupby('town').agg(['min', 'max'])['unit_price'].reset_index()
town_ave.columns = ['town', 'town_min_unit_price', 'town_max_unit_price']

# Only merged if column not exist - to ease re-running of code
if 'town_min_unit_price' not in hex_df.columns:
    hex_df = pd.merge(hex_df, town_ave, how='left', on='town')

# Normalize price of property with respect to its HDB Township
max_ = hex_df['town_max_unit_price']
min_ = hex_df['town_min_unit_price']
hex_df['appeal_score'] = (hex_df['unit_price'] - min_) / ( max_ - min_)

## 5. Final Data Export

hex_df.to_csv('./Dataset/final_data_with_hex.csv', index=False)

Finally we are ready for the analysis and modelling. It will be continued in the last notebook, Part 7.
